+++ server/utils/songsControllerUtils.js
@@ -1,3 +1,22 @@
+const createNewSongObject = (scrapedData) => {
+    const {name, lyrics, artist, originalLang } = scrapedData.body
+    const newSongObject = {
+        name: {
+        hebrew: name.hebrew,
+         arabic:name.arabic,
+          english:name.english.toLowerCase()
+        },
+      lyrics: {
+        hebrew: lyrics.hebrew,
+         arabic:lyrics.arabic,
+          english:lyrics.english
+        },
+      originalLang,
+      artist
+    }
+    return newSongObject
+}
+
 
 const createSearchFilterObject = (req) => {
     const {name, album, artist} = req.body
@@ -19,4 +38,5 @@ const createSearchFilterObject = (req) => {
     return filter
 }
 
-export default createSearchFilterObject
+
+export {createNewSongObject, createSearchFilterObject}
